Update Modules names
