+++ app/build.gradle
@@ -47,10 +47,9 @@ android {
 }
 
 dependencies {
-
     //Modules
-    implementation project(':podcast_details')
     implementation project(':core')
+    implementation project(':data')
 
     def composeBom = platform('androidx.compose:compose-bom:2023.01.00')
     implementation composeBom
