+++ app/src/main/java/com/example/podcastapp/App.kt
@@ -1,8 +1,8 @@
 package com.example.podcastapp
 
 import android.app.Application
-import com.example.podcast_details.di.repositoryModule
-import com.example.podcastapp_core.di.apiModule
+import com.example.data.di.repositoryModule
+import com.example.core.di.apiModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
