+++ core/src/main/graphql/com/example/core/schema.sdl
