+++ data/src/main/AndroidManifest.xml
