+++ data/src/main/graphql/com/example/data/PodcastDetails.graphql
