+++ data/src/main/graphql/com/example/data/SearchForTerm.graphql
@@ -1,0 +1,51 @@
+query SearchAll($term: String!,$filterForTypes:[TaddyType]!, $limitPerPage: Int!) {
+    searchForTerm(term:$term, filterForTypes:$filterForTypes, limitPerPage: $limitPerPage){
+        searchId
+        podcastSeries{
+            uuid
+            name
+            imageUrl
+            description
+        }
+        podcastEpisodes{
+            uuid
+            name
+            audioUrl
+            subtitle
+            duration
+            podcastSeries{
+                uuid
+                name
+            }
+        }
+    }
+}
+
+query SearchSeriesForTerm($term: String!, $limitPerPage: Int!) {
+    searchForTerm(term:$term, filterForTypes:PODCASTSERIES, limitPerPage: $limitPerPage){
+        podcastSeries{
+            uuid
+            name
+            imageUrl
+            description
+        }
+    }
+}
+
+
+
+query SearchEpisodesForTerm($term: String!, $limitPerPage: Int!) {
+    searchForTerm(term:$term, filterForTypes:PODCASTEPISODE, limitPerPage: $limitPerPage){
+        podcastEpisodes{
+            uuid
+            name
+            audioUrl
+            subtitle
+            duration
+            podcastSeries{
+                uuid
+                name
+            }
+        }
+    }
+}
