+++ data/src/main/java/com/example/data/di/RepositoriesModule.kt
@@ -1,0 +1,11 @@
+package com.example.data.di
+
+import com.example.data.repositories.IPodcastRepository
+import com.example.data.repositories.PodcastRepository
+import org.koin.dsl.module
+
+val repositoryModule = module {
+    single<IPodcastRepository> {
+        PodcastRepository(get(), get())
+    }
+}
