--- podcast_details/src/main/java/com/example/podcast_details/di/RepositoriesModule.kt
@@ -1,13 +1,0 @@
-package com.example.podcast_details.di
-
-import com.example.podcast_details.data.repositories.IPodcastRepository
-import com.example.podcast_details.data.repositories.PodcastRepository
-import org.koin.dsl.module
-
-val repositoryModule = module {
-    single<IPodcastRepository> {
-        PodcastRepository(
-            get()
-        )
-    }
-}
