+++ settings.gradle
@@ -14,5 +14,5 @@ dependencyResolutionManagement {
 }
 rootProject.name = "PodcastApp"
 include ':app'
-include ':podcast_details'
+include ':data'
 include ':core'
